requirements.txt
black==23.3.0
click==8.1.3
exceptiongroup==1.1.1
Faker==18.4.0
gevent==22.10.2
greenlet==2.0.2
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.3.0
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
zope.event==4.6
zope.interface==6.0
